# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DriveTrack - Driving School Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /moderation/cases:
    get:
      operationId: getModerationCases
      tags: [moderation]
      summary: List moderation cases (super_admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: schoolId
          in: query
          schema:
            type: integer
        - name: contentType
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Cases list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModerationCase"
  /moderation/cases/{id}:
    get:
      operationId: getModerationCase
      tags: [moderation]
      summary: Get a moderation case with its event history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationCaseDetail"
    patch:
      operationId: updateModerationCase
      tags: [moderation]
      summary: Update case status, legalHold, or reviewOutcome
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationCasePatch"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationCase"
  /moderation/cases/{id}/release:
    post:
      operationId: releaseModerationCase
      tags: [moderation]
      summary: Release quarantined content after review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                outcome:
                  type: string
      responses:
        "200":
          description: Case released
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /moderation/cases/{id}/escalate:
    post:
      operationId: escalateModerationCase
      tags: [moderation]
      summary: Escalate a moderation case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Case escalated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /moderation/exports:
    post:
      operationId: createModerationExport
      tags: [moderation]
      summary: Create a law enforcement evidence export (super_admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationExportInput"
      responses:
        "201":
          description: Export created
          content:
            application/json:
              schema:
                type: object
                required: [exportId, caseCount]
                properties:
                  exportId:
                    type: integer
                  caseCount:
                    type: integer

  # ─── Billing ──────────────────────────────────────────────────────────────────
...
  /demo/reset:
    post:
      operationId: resetDemoData
      tags: [demo]
      summary: Reset demo school data (super_admin only, feature-flagged)
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoResetInput"
      responses:
        "200":
          description: Demo data reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoResetResult"
components:
  schemas:
...
    ModerationCase:
      type: object
      required: [id, status, severity, contentType]
      properties:
        id:
          type: integer
        schoolId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [open, under_review, escalated, released, closed]
        severity:
          type: string
          enum: [low, medium, high, critical]
        contentType:
          type: string
        contentId:
          type: ["integer", "null"]
        actorUserId:
          type: ["integer", "null"]
        targetUserId:
          type: ["integer", "null"]
        studentId:
          type: ["integer", "null"]
        rawExcerpt:
          type: ["string", "null"]
        ruleHitsJson:
          type: ["object", "null"]
        reviewOutcome:
          type: ["string", "null"]
        reviewedByUserId:
          type: ["integer", "null"]
        reviewedAt:
          type: ["string", "null"]
        legalHold:
          type: boolean
        createdAt:
          type: string
    ModerationCasePatch:
      type: object
      properties:
        status:
          type: string
          enum: [open, under_review, escalated, released, closed]
        reviewOutcome:
          type: string
        legalHold:
          type: boolean
    ModerationEvent:
      type: object
      required: [id, moderationCaseId, eventType]
      properties:
        id:
          type: integer
        moderationCaseId:
          type: integer
        eventType:
          type: string
          enum: [detected, quarantined, reviewed, released, escalated, exported, closed]
        payloadJson:
          type: ["object", "null"]
        createdAt:
          type: string

    # ─── Subscriptions ────────────────────────────────────────────────────────
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        message:
          type: ["string", "null"]

    # ─── Schools ──────────────────────────────────────────────────────────────
...
    ModerationCaseDetail:
      allOf:
        - $ref: "#/components/schemas/ModerationCase"
        - type: object
          properties:
            events:
              type: array
              items:
                $ref: "#/components/schemas/ModerationEvent"
    ModerationExportInput:
      type: object
      required: [caseIds, reason]
      properties:
        caseIds:
          type: array
          items:
            type: integer
        reason:
          type: string
        schoolId:
          type: integer
...
    DemoResetInput:
      type: object
      properties:
        resetScope:
          type: string
          enum: [full_demo, bookings_only, students_only]
        notes:
          type: string
    DemoResetResult:
      type: object
      required: [ok, scope]
      properties:
        ok:
          type: boolean
        resetId:
          type: integer
        scope:
          type: string